+++ test/maps/vertexmap.jl
@@ -6,22 +6,30 @@ vm = ConstVertexMap(0)
 @test get(vm, -1, 1) == 0
 vm[1] = 1
 @test vm[1] == 0
-@test hasindex(vm, 1)
-@test hasindex(vm, -1)
+@test haskey(vm, 1)
+@test haskey(vm, -1)
 @test length(vm) == typemax(Int)
 
-vm = rand(1:10,10)
-@test typeof(vm) <: AVertexMap
-@test hasindex(vm, 1)
-@test !hasindex(vm, -1)
-
 g = G()
-vm = Dict{V, Float64}()
-@test VertexMap(g, Float64) == vm
+vm = VertexMap(g, rand(1:10,10))
+@test typeof(vm) <: AVertexMap
+@test haskey(vm, 2)
+@test !haskey(vm, -1)
+@test valtype(vm) == Int
+@test get(vm, 1, -100) != -100
+
+vm = VertexMap(g, Float64)
+@test valtype(vm) == Float64
 @test typeof(vm) <: AVertexMap
 
 vm[1] = 2.
-@test hasindex(vm, 1)
-@test !hasindex(vm, 2)
+@test haskey(vm, 1)
+@test !haskey(vm, 2)
+@test get(vm, -1, -100) == -100
+
+vm2 = deepcopy(vm)
+@test vm2 == vm
+
+@test sprint(show, vm) == "VertexMap: $(vm.data)"
 
 end # testset